requirements.txt
customtkinter
ttkwidgets
Pillow
tiktoken